requirements.txt
# --------------------
# Core Python & Data
# --------------------
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2
scipy==1.11.4


# --------------------
# NLP (Traditional)
# --------------------
nltk
spacy

# --------------------
# Speech Recognition
# --------------------
SpeechRecognition
pydub
numba==0.59.1
llvmlite==0.42.0
openai-whisper


# --------------------
# LLM + RAG
# --------------------
sentence-transformers==5.2.2
faiss-cpu==1.13.2
torch==2.10.0
transformers==5.1.0

# --------------------
# Utilities
# --------------------
tqdm
python-dotenv
regex
